# api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Optimus AI Platform API",
    "description": "Full REST API specification for the Optimus AI platform covering authentication, users, chats, models, files, knowledge, prompts, tools, groups, functions, skills, memories, notes, channels, automations, calendar, folders, configs, and third-party integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "Optimus AI Support",
      "url": "https://github.com/optimus-ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/users/": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user",
        "operationId": "getCurrentUser",
        "responses": {
          "200": { "description": "User info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users",
        "operationId": "searchUsers",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Search query" }
        ],
        "responses": {
          "200": { "description": "Users found", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } }
        }
      }
    },
...
    "/api/v1/users/default/permissions": {
      "get": {
        "tags": ["Users"],
        "summary": "Get default permissions",
        "operationId": "getDefaultPermissions",
        "responses": {
          "200": { "description": "Default permissions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Permissions" } } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Update default permissions",
        "operationId": "updateDefaultPermissions",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Permissions" } } }
        },
        "responses": {
          "200": { "description": "Permissions updated" }
        }
      }
    },
...
    "/api/v1/users/user/settings": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user settings",
        "operationId": "getUserSettings",
        "responses": {
          "200": { "description": "User settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserSettings" } } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Update user settings",
        "operationId": "updateUserSettings",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserSettings" } } }
        },
        "responses": {
          "200": { "description": "Settings updated" }
        }
      }
    },
    "/api/v1/users/user/info": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user info",
        "operationId": "getUserInfo",
        "responses": {
          "200": { "description": "User info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Update user info",
        "operationId": "updateUserInfo",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
        },
        "responses": {
          "200": { "description": "Info updated" }
        }
      }
    },
    "/api/v1/users/{userId}/update": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "operationId": "getUserById",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Update user by ID",
        "operationId": "updateUserById",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
        },
        "responses": {
          "200": { "description": "User updated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID",
        "operationId": "deleteUserById",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/chats/list/user/{userId}": {
      "get": {
        "tags": ["Chats"],
        "summary": "List chats for a user",
        "operationId": "listChatsByUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User's chats", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Chat" } } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtained via /api/v1/auths/signin or /api/v1/auths/signup."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key with 'sk-' prefix. Passed as 'Bearer sk-...'"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized – missing or invalid credentials",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string", "description": "Error message" }
        },
        "required": ["detail"]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "user"] },
          "status": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "profileImage": { "type": "string", "format": "uri" }
        },
        "required": ["id", "name", "email", "role"]
      },
...
      "Permissions": {
        "type": "object",
        "properties": {
          "models": { "type": "boolean" },
          "prompts": { "type": "boolean" },
          "files": { "type": "boolean" },
          "knowledge": { "type": "boolean" },
          "tools": { "type": "boolean" },
          "chats": { "type": "boolean" }
        }
      },
...
      "UserSettings": {
        "type": "object",
        "properties": {
          "theme": { "type": "string", "enum": ["light", "dark", "system"] },
          "language": { "type": "string" },
          "notifications": { "type": "boolean" },
          "defaultModel": { "type": "string" }
        }
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "role": { "type": "string", "enum": ["system", "user", "assistant", "tool"] },
          "content": { "type": ["string", "null"] },
          "name": { "type": "string" },
          "tool_calls": { "type": "array", "items": { "$ref": "#/components/schemas/ToolCall" } },
          "tool_call_id": { "type": "string" }
        },
        "required": ["role"]
      },
      "ToolCall": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "default": "function" },
          "function": { "$ref": "#/components/schemas/FunctionCall" }
        }
      },
      "FunctionCall": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "arguments": { "type": "string" }
        }
      },
...
      "Chat": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "chatId": { "type": "string" },
          "userId": { "type": "string" },
          "models": { "type": "array", "items": { "type": "string" } },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" } },
          "history": { "type": "object", "additionalProperties": {} },
          "tags": { "type": "array", "items": { "type": "string" } },
          "pinned": { "type": "boolean" },
          "archived": { "type": "boolean" },
          "shared": { "type": "boolean" },
          "shareId": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}